add edit functionality

+++ src/lib/components/ResourceCard/ResourceCard.svelte
@@ -3,13 +3,18 @@
 	import * as Card from '$lib/components/ui/card';
 	import CardType from '$lib/components/ResourceCard/CardType.svelte';
 	import Time from 'svelte-time/Time.svelte';
-	import { CircleUser, ExternalLink, Pencil, Tag, Trash2, User } from '@lucide/svelte';
+	import { CircleUser, Edit, ExternalLink, Pencil, Trash2 } from '@lucide/svelte';
 	import { drawer } from '$lib/components/Drawer/drawerState.svelte';
-	import type { Resource } from '$lib/supabase/resource.types';
+	import type { Resource } from '$lib/supabase/schema.types';
 
-	let { resource }: { resource: Resource } = $props();
+	let { resource, currentUserId }: { resource: Resource; currentUserId: string } = $props();
 
-	const handleEdit = (event: MouseEvent) => event.stopPropagation();
+	const handleEdit = (event: MouseEvent) => {
+		event.stopPropagation();
+		drawer.action = 'edit';
+		drawer.isOpen = true;
+		drawer.resource = resource;
+	};
 
 	const handleDelete = (event: MouseEvent) => event.stopPropagation();
 </script>
@@ -17,6 +22,7 @@
 <Card.Root
 	class="flex cursor-pointer flex-col justify-between gap-2.5"
 	onclick={() => {
+		drawer.action = 'view';
 		drawer.isOpen = true;
 		drawer.resource = resource;
 	}}
@@ -25,17 +31,19 @@
 		<CardType type={resource.type} />
 		<Card.Title class="text-lg">{resource.title}</Card.Title>
 		<Card.Description class="flex-1 text-primary">{resource.description}</Card.Description>
-		<Card.Action class="flex gap-2.5">
-			<button onclick={(event) => handleEdit(event)}>
-				<Pencil class="cursor-pointer" size={20} />
-			</button>
-			<form method="POST" action="?/delete">
-				<button onclick={(event) => handleDelete(event)}>
-					<Trash2 class="cursor-pointer" size={20} />
-					<input type="hidden" name="id" value={resource.id} />
+		{#if resource?.author_id === currentUserId}
+			<Card.Action class="flex gap-2.5">
+				<button type="button" onclick={(event: MouseEvent) => handleEdit(event)}>
+					<Pencil class="cursor-pointer" size={20} />
 				</button>
-			</form>
-		</Card.Action>
+				<form method="POST" action="?/delete">
+					<button onclick={(event) => handleDelete(event)}>
+						<Trash2 class="cursor-pointer" size={20} />
+						<input type="hidden" name="id" value={resource.id} />
+					</button>
+				</form>
+			</Card.Action>
+		{/if}
 	</Card.Header>
 	<Card.Action class="px-6 pb-1">
 		<a class="flex items-center gap-2" href={resource.url} target="_blank">
